+++ methodical-jackson2/src/test/java/org/jwcarman/methodical/jackson2/Jackson2ParameterResolverTest.java
@@ -28,7 +28,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.jwcarman.methodical.MethodInvocationException;
+import org.jwcarman.methodical.ParameterResolutionException;
 import org.jwcarman.methodical.param.ParameterInfo;
 
 class Jackson2ParameterResolverTest {
@@ -104,7 +104,7 @@ class Jackson2ParameterResolverTest {
     ParameterInfo info = paramInfo("value", 0, int.class);
     JsonNode params = mapper.readTree("{\"value\": \"not a number\"}");
     assertThatThrownBy(() -> resolver.resolve(info, params))
-        .isInstanceOf(MethodInvocationException.class)
+        .isInstanceOf(ParameterResolutionException.class)
         .hasMessageContaining("Unable to deserialize parameter");
   }